Remove unnecessary rest configs
